adapter_1.1

+++ app/src/main/java/com/example/dipnetocom/dto/FeedItem.kt
@@ -1,5 +1,5 @@
 package com.example.dipnetocom.dto
 
 sealed class FeedItem {
-    abstract val id: Long
+    abstract val id: Int
 }
